✨ Add clubs, add matches, add more stats grouped by seasons

+++ services/api/src/graphql/stats/object-types/stat-by-season.ts
@@ -1,11 +1,11 @@
-import { Field, Int, ObjectType } from 'type-graphql'
+import { Field, Float, ObjectType } from 'type-graphql'
 
 @ObjectType()
 export class StatBySeason {
     @Field({ description: 'Season the stat is calculated for.' })
     season: string
 
-    @Field(() => Int, {
+    @Field(() => Float, {
         description: 'Result value of a calculated stat.',
     })
     value: number
